Use same kernal name as TF C++ version
