+++ tfjs-core/src/backends/cpu/all_kernels.ts
@@ -19,4 +19,4 @@
 // global registry when we compile the library. A modular build would replace
 // the contents of this file and import only the kernels that are needed.
 import './square';
-import './non_max_suppression_with_score';
+import './non_max_suppression_v5';
