+++ tfjs-core/src/backends/webgl/all_kernels.ts
@@ -20,4 +20,4 @@
 // the contents of this file and import only the kernels that are needed.
 import './square';
 import './fromPixels';
-import './non_max_suppression_with_score';
+import './non_max_suppression_v5';
